separated shaky tests into separate system_tests_long_shaky test suite (test_collapse, test_concurrent_collapse_fails, test_catchup_exercises, test_restart_master_long, test_243, test_272, test_inject_as_head)
