+++ jenkins/020_update_qpackages.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -xue
+
+/opt/qbase3/qshell -c "
+i.qp.updateAll()
+"
