+++ jenkins/030_prepare_system_tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -xue
+
+/opt/qbase3/qshell ./jenkins/prepareSystemTests.py
+
+mkdir -p ${WORKSPACE}/testresults
