+++ jenkins/050_run_pending_configuration.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -xue
+
+/opt/qbase3/qshell -c "
+q.qp._runPendingReconfigeFiles()
+"
