+++ jenkins/seed
@@ -1,0 +1,1 @@
+arakoon-1.x-qbase3-ubuntu-12.10-quantal-20130514
