+++ jenkins/system_tests_slow_right/020_update_qpackages.sh
@@ -1,5 +1,1 @@
-#!/bin/bash -xue
-
-/opt/qbase3/qshell -c "
-i.qp.updateAll()
-"
+../020_update_qpackages.sh
