+++ jenkins/system_tests_slow_right/050_run_pending_configuration.sh
@@ -1,5 +1,1 @@
-#!/bin/bash -xue
-
-/opt/qbase3/qshell -c "
-q.qp._runPendingReconfigeFiles()
-"
+../050_run_pending_configuration.sh
