+++ jenkins/system_tests_slow_right/seed
@@ -1,1 +1,1 @@
-arakoon-1.x-qbase3-ubuntu-12.10-quantal-20130514
+../seed
