+++ jenkins/system_tests_slow_shaky/060_install_nose.sh
