+++ jenkins/system_tests_slow_shaky/080_archive_version.sh
@@ -1,0 +1,1 @@
+../archive_version.sh
